Set known defaults for a `CellModel`s `ODESystem`
CellML files contain `initial_value` fields specifying initial values
for parameters and states in the system. Retrieving these for the
`ODESystem` was already possible, but had to be done separately or as
part of the `ODEProblem` constructor defined for `CellModel`s.

Defaults are only set for simplified `ODESystem`s. Unsimplified
`ODESystem`s may contain parameters and states for which no defaults are
available resulting in errors trying to construct the `ODESystem`. These
errors will still occur when using the result of calling
`process_components` with `simplify = false`, but as noted in the code
that should only be done for debugging purposes and this functionality
is not exposed at the higher API levels.

Fixes #56.